+++ openpyxl2/packaging/manifest.py
@@ -212,6 +212,9 @@ def write_content_types(workbook, as_template=False, exts=None):
             name = '/xl/comments%d.xml' % comments_id
             manifest.Override.append(Override(name, COMMENTS_TYPE))
 
+        for t in sheet._tables:
+            manifest.Override.append(Override(t.abs_path, t._type))
+
 
     # chartsheets
     for sheet in workbook.chartsheets: