+++ openpyxl2/worksheet/header_footer.py
@@ -17,8 +17,8 @@ from openpyxl2.descriptors import (
 from openpyxl2.descriptors.serialisable import Serialisable
 
 from openpyxl2.xml.functions import Element
+from openpyxl2.utils.escape import escape, unescape
 
-RGB = ("^[A-Fa-f0-9]{6}$")
 
 FONT_PATTERN = '&"(?P<font>.+)"'
 COLOR_PATTERN  = "&K(?P<color>[A-F0-9]{6})"
@@ -27,19 +27,19 @@ FORMAT_REGEX = re.compile("{0}|{1}|{2}".format(FONT_PATTERN, COLOR_PATTERN,
                                                SIZE_REGEX)
                           )
 
-# See http://stackoverflow.com/questions/27711175/regex-with-multiple-optional-groups for discussion
-ITEM_REGEX = re.compile("""
-(&L(?P<left>.+?))?
-(&C(?P<center>.+?))?
-(&R(?P<right>.+?))?
-$""", re.VERBOSE | re.DOTALL)
-
-# add support for multiline strings (how do re.flags combine?)
-
 def _split_string(text):
     """
     Split the combined (decoded) string into left, center and right parts
+
+    # See http://stackoverflow.com/questions/27711175/regex-with-multiple-optional-groups for discussion
     """
+
+    ITEM_REGEX = re.compile("""
+    (&L(?P<left>.+?))?
+    (&C(?P<center>.+?))?
+    (&R(?P<right>.+?))?
+    $""", re.VERBOSE | re.DOTALL)
+
     m = ITEM_REGEX.match(text)
     try:
         parts = m.groupdict()
@@ -62,8 +62,7 @@ class _HeaderFooterPart(Strict):
     * &B   Toggles bold
     * &D or &[Date]   Inserts the current date
     * &E   Toggles double-underline
-    * &F or &[File]   Inserts the workbook name    oddHeader = Typed(expected_type=HeaderFooterItem)
-    oddFooter = Typed(expected_type=HeaderFooterItem)
+    * &F or &[File]   Inserts the workbook name
     * &I   Toggles italic
     * &N or &[Pages]   Inserts the total page count
     * &S   Toggles strikethrough
@@ -86,6 +85,7 @@ class _HeaderFooterPart(Strict):
     text = String(allow_none=True)
     font = String(allow_none=True)
     size = Integer(allow_none=True)
+    RGB = ("^[A-Fa-f0-9]{6}$")
     color = MatchPattern(allow_none=True, pattern=RGB)
 
 
@@ -159,13 +159,30 @@ class HeaderFooterItem(Strict):
         """
         Pack parts into a single string
         """
+        TRANSFORM = {'&[Tab]': '&A', '&[Pages]': '&N', '&[Date]': '&D',
+                     '&[Path]': '&Z', '&[Page]': '&P', '&[Time]': '&T', '&[File]': '&F',
+                     '&[Picture]': '&G'}
+
+        # escape keys and create regex
+        SUBS_REGEX = re.compile("|".join(["({0})".format(re.escape(k))
+                                          for k in TRANSFORM]))
+
+        def replace(match):
+            """
+            Callback for re.sub
+            Replace expanded control with mini-format equivalent
+            """
+            sub = match.group(0)
+            return TRANSFORM[sub]
+
         txt = []
         for key, part in zip(
             self.__keys, [self.left, self.center, self.right]):
             if part.text is not None:
                 txt.append("&{0}{1}".format(key, str(part)))
         txt = "".join(txt)
-        return SUBS_REGEX.sub(replace, txt)
+        txt = SUBS_REGEX.sub(replace, txt)
+        return escape(txt)
 
 
     def __bool__(self):
@@ -186,32 +203,13 @@ class HeaderFooterItem(Strict):
     @classmethod
     def from_tree(cls, node):
         if node.text:
-            parts = _split_string(node.text)
+            text = unescape(node.text)
+            parts = _split_string(text)
             for k, v in parts.items():
                 if v is not None:
                     parts[k] = _HeaderFooterPart.from_str(v)
             self = cls(**parts)
             return self
-
-
-
-TRANSFORM = {'&[Tab]': '&A', '&[Pages]': '&N', '&[Date]': '&D', '\n': '_x000D_',
-        '&[Path]': '&Z', '&[Page]': '&P', '&[Time]': '&T', '&[File]': '&F',
-        '&[Picture]': '&G'}
-
-
-# escape keys and create regex
-SUBS_REGEX = re.compile("|".join(["({0})".format(re.escape(k))
-                                  for k in TRANSFORM]))
-
-
-def replace(match):
-    """
-    Callback for re.sub
-    Replace expanded control with mini-format equivalent
-    """
-    sub = match.group(0)
-    return TRANSFORM[sub]
 
 
 class HeaderFooter(Serialisable):