+++ openpyxl2/worksheet/tests/test_header.py
@@ -8,13 +8,15 @@ from openpyxl2.tests.helper import compare_xml
 
 
 def test_split_into_parts():
-    from .. header_footer import ITEM_REGEX
-    m = ITEM_REGEX.match("&Ltest header")
-    assert m.group('left') == "test header"
-    m = ITEM_REGEX.match("""&L&"Lucida Grande,Standard"&K000000Left top&C&"Lucida Grande,Standard"&K000000Middle top&R&"Lucida Grande,Standard"&K000000Right top""")
-    assert m.group('left') == '&"Lucida Grande,Standard"&K000000Left top'
-    assert m.group('center') == '&"Lucida Grande,Standard"&K000000Middle top'
-    assert m.group('right') == '&"Lucida Grande,Standard"&K000000Right top'
+    from .. header_footer import _split_string
+
+    headers = _split_string("&Ltest header")
+    assert headers['left'] == "test header"
+
+    headers = _split_string("""&L&"Lucida Grande,Standard"&K000000Left top&C&"Lucida Grande,Standard"&K000000Middle top&R&"Lucida Grande,Standard"&K000000Right top""")
+    assert headers['left'] == '&"Lucida Grande,Standard"&K000000Left top'
+    assert headers['center'] == '&"Lucida Grande,Standard"&K000000Middle top'
+    assert headers['right'] == '&"Lucida Grande,Standard"&K000000Right top'
 
 
 def test_cannot_split():
@@ -25,21 +27,29 @@ def test_cannot_split():
 
 
 def test_multiline_string():
-    from .. header_footer import ITEM_REGEX
+    from .. header_footer import _split_string
+
     s = """&L141023 V1&CRoute - Malls\nSchedules R1201 v R1301&RClient-internal use only"""
-    match = ITEM_REGEX.match(s)
-    assert match.groupdict() == {
+    headers = _split_string(s)
+    assert headers == {
         'center': 'Route - Malls\nSchedules R1201 v R1301',
         'left': '141023 V1',
         'right': 'Client-internal use only'
     }
 
 
-def test_font_size():
-    from .. header_footer import SIZE_REGEX
-    s = "&9"
-    match = re.search(SIZE_REGEX, s)
-    assert match.group('size') == "9"
+@pytest.mark.parametrize("value, expected",
+                         [
+                             ("&9", [('', '', '9')]),
+                             ('&"Lucida Grande,Standard"', [("Lucida Grande,Standard", '', '')]),
+                             ('&K000000', [('', '000000', '')])
+                         ]
+                         )
+def test_parse_format(value, expected):
+    from .. header_footer import FORMAT_REGEX
+
+    m = FORMAT_REGEX.findall(value)
+    assert m == expected
 
 
 @pytest.fixture
@@ -69,13 +79,6 @@ class TestHeaderFooterPart:
         assert bool(hf) is False
         hf.text = "Title"
         assert bool(hf) is True
-
-
-def test_subs():
-    from ..header_footer import SUBS_REGEX, replace
-    s = "MyName&[Tab]&[Page]&[Path]"
-    t = SUBS_REGEX.sub(replace, s)
-    assert t == "MyName&A&P&Z"
 
 
 @pytest.fixture