+++ doc/changes.rst
@@ -28,6 +28,12 @@ raises when inspecting EmptyCells in read-only mode
 * `629`<https://bitbucket.org/openpyxl/openpyxl/issues/629>`_ Row Dimensions not supported in write-only mode
 
 
+Features
+--------
+
+* Add write support for worksheet tables
+
+
 2.4.0-a1 (2016-04-11)
 =====================